More of partial Impbook backend  (#154)
* more of import books backend, sort of doing TDD which is quite nice actually

* Import book Domain service added, most of update implemented, or facade side of things anyway

* changing to accept list of sub objects, started implementing domain service more, started trying to write tests for domain service

* some more domain service, unsure if will work

* pr work - adding missing cog of service db context link between parent and child tables, rewrite domain service and new tests on it

* changed to ILists so can mutate the data, got tests passing for add and update but not still need to change things for remove list items

* deleted code for removing list elements in domain service, added tests for post entries and order details

* move import book keys from parts folder to import books folder + move constructors to top

* code formatting

* more code formatting

* stylecop complaints fixed and use of decimal cast changed to m

* remove accidentally commited stylecop settings

* facade test, can't get to work

* formatted test file

* don't update datecreated and change tests to reflect that

* facade tests fixed

+++ src/Resources/ImportBookInvoiceDetailResource.cs
@@ -10,6 +10,6 @@
 
         public string InvoiceNumber { get; set; }
 
-        public int InvoiceValue { get; set; }
+        public decimal InvoiceValue { get; set; }
     }
 }
